[fix] Fix matrix transformation in geometry shader
ジオメトリシェーダーで行列変換するように修正

+++ Buffer.cpp
@@ -68,7 +68,7 @@ IndexBuffer* IndexBuffer::Create(ID3D11Device* device, UINT indexCount)
 
 	// インデックスバッファーについての記述
 	D3D11_BUFFER_DESC bufferDesc = {};
-	bufferDesc.ByteWidth = sizeof(UINT16) * indexCount;
+	bufferDesc.ByteWidth = sizeof(UINT32) * indexCount;
 	bufferDesc.Usage = D3D11_USAGE_DEFAULT;
 	bufferDesc.BindFlags = D3D11_BIND_INDEX_BUFFER;
 	bufferDesc.CPUAccessFlags = 0;
@@ -84,7 +84,7 @@ IndexBuffer* IndexBuffer::Create(ID3D11Device* device, UINT indexCount)
 }
 
 // バッファーにデータを設定する
-void IndexBuffer::SetData(UINT16* data)
+void IndexBuffer::SetData(UINT32* data)
 {
 	// デバイスの取得
 	ComPtr<ID3D11Device> device = nullptr;
